+++ R/function_PlotMapPoints.R
@@ -39,9 +39,9 @@
 #' @param col.labels A character vector, specifying custom labels to be used for each legend item. Works with \code{map.type} set to \code{default} or \code{leaflet}.
 #' @param col.rev Logical, If \code{TRUE}, then color palette will be reversed.
 #' @param plot.scale Logical, plot a scale bar on map. NOTE: Scale bar may be inaccurate for geographic coordinate systems (Consider switching to projected coordinate system).
-#' @param scale.pos Keyword string for scalebar position for static maps. One of \code{bl}, \code{br}, \code{tr}, or \code{tl}. See \code{\link{annotation_scale}}.
+#' @param scale.pos Keyword string for scalebar position for static maps. One of \code{bl}, \code{br}, \code{tr}, or \code{tl}.
 #' @param plot.arrow Logical, plot a North arrow in static maps.
-#' @param arrow.pos Keyword string for north arrow position for static maps. One of \code{bl}, \code{br}, \code{tr}, or \code{tl}. See \code{\link{annotation_north_arrow}}.
+#' @param arrow.pos Keyword string for north arrow position for static maps. One of \code{bl}, \code{br}, \code{tr}, or \code{tl}.
 #' @param radius Numeric, radius of markers maps. See \code{\link{geom_sf}} for static maps and [leaflet::addCircleMarkers()] for Leaflet maps.
 #' @param weight Numeric, weight of marker outlines in Leaflet maps. See [leaflet::addCircleMarkers()].
 #' @param opacity Numeric, opacity of marker outlines in Leaflet maps. See [leaflet::addCircleMarkers()].
@@ -120,7 +120,6 @@
 #' 
 #' @importFrom dplyr right_join %>% mutate filter across
 #' @importFrom ggplot2 aes geom_sf ggplot ggsave scale_color_manual scale_fill_manual theme element_text element_blank
-#' @importFrom ggspatial annotation_north_arrow annotation_scale
 #' @importFrom grDevices dev.list colorRampPalette
 #' @importFrom graphics par frame legend strwidth text plot.new
 #' @importFrom stats setNames
@@ -824,13 +823,13 @@ PlotMapPoints <- function(x, sites, sites.subid.column = 1, sites.groups = NULL,
         # Add scale bar
         if(plot.scale == TRUE){
           plot <- plot +
-            annotation_scale(location = scale.pos)
+            .annotation_scale(location = scale.pos)
         }
         
         # Add north arrow
         if(plot.arrow == TRUE){
           plot <- plot +
-            annotation_north_arrow(location = arrow.pos)
+            .annotation_north_arrow(location = arrow.pos)
         }
         
         # Save image